# Dryft/backend/openapi.yaml
openapi: 3.0.3
info:
  title: Drift API
  description: Backend API for the Drift social/dating application.
  version: 1.0.0
  contact:
    name: Drift Team
servers:
  - url: https://api.drift.app
    description: Production
  - url: http://localhost:8080
    description: Local development
...
security:
  - BearerAuth: []
paths:
...
  /v1/age-gate/status:
    get:
      tags: [Age Verification]
      summary: Get current verification status
      operationId: getVerificationStatus
      responses:
        "200":
          description: Verification status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationStatus"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /v1/admin/verifications/pending:
    get:
      tags: [Admin]
      summary: Get pending verifications
      operationId: getPendingVerifications
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Pending verifications
          content:
            application/json:
              schema:
                type: object
                properties:
                  verifications:
                    type: array
                    items:
                      $ref: "#/components/schemas/VerificationRecord"
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        default: 0
  responses:
...
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
...
    VerificationStatus:
      type: object
      properties:
        status:
          type: string
          enum: [none, pending, approved, rejected]
        method:
          type: string
          enum: [card, id]
        verified_at:
          type: string
          format: date-time
          nullable: true
...
    VerificationRecord:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        method:
          type: string
        status:
          type: string
        submitted_at:
          type: string
          format: date-time
        reviewed_at:
          type: string
          format: date-time
          nullable: true
        reviewer_notes:
          type: string
...